// packages/payload/src/fields/config/types.ts
export type FieldHook<T extends TypeWithID = any, P = any, S = any> = (
  args: FieldHookArgs<T, P, S>,
) => P | Promise<P>

export type FieldAccess<T extends TypeWithID = any, P = any, U = any> = (args: {
  /**
   * The incoming data used to `create` or `update` the document with. `data` is undefined during the `read` operation.
   */
  data?: Partial<T>
  /**
   * The original data of the document before the `update` is applied. `doc` is undefined during the `create` operation.
   */
  doc?: T
  /**
   * The `id` of the current document being read or updated. `id` is undefined during the `create` operation.
   */
  id?: number | string
  /** The `Express` request object containing the currently authenticated `user` */
  req: PayloadRequest<U>
  /**
   * Immediately adjacent data to this field. For example, if this is a `group` field, then `siblingData` will be the other fields within the group.
   */
  siblingData?: Partial<P>
}) => Promise<boolean> | boolean

export type Condition<T extends TypeWithID = any, P = any> = (
  data: Partial<T>,
  siblingData: Partial<P>,
  { user }: { user: User },
) => boolean

export type FilterOptionsProps<T = any> = {
  data: T
  id: number | string
  relationTo: string
  siblingData: unknown
  user: Partial<User>
}

export type FilterOptions<T = any> =
  | ((options: FilterOptionsProps<T>) => Promise<Where | boolean> | Where | boolean)
  | Where
  | null

type Admin = {
  className?: string
  components?: {
    Cell?: React.ComponentType<any>
    Field?: React.ComponentType<any>
    Filter?: React.ComponentType<any>
  }
  /**
   * You can programmatically show / hide fields based on what other fields are doing.
   * This is also run on the server, to determine if the field should be validated.
   */
  condition?: Condition
  description?: Description
  disableBulkEdit?: boolean
  disabled?: boolean
  hidden?: boolean
  position?: 'sidebar'
  readOnly?: boolean
  style?: CSSProperties
  width?: string
}
// ...
export type Validate<TValue = any, TData = any, TSiblingData = any, TFieldConfig = any> = (
  value: TValue,
  options: ValidateOptions<TData, TSiblingData, TFieldConfig>,
) => Promise<string | true> | string | true

export type OptionObject = {
  label: Record<string, string> | string
  value: string
}

export type Option = OptionObject | string

export interface FieldBase {
  access?: {
    create?: FieldAccess
    read?: FieldAccess
    update?: FieldAccess
  }
  admin?: Admin
  /** Extension point to add your custom data. */
  custom?: Record<string, any>
  defaultValue?: any
  hidden?: boolean
  hooks?: {
    afterChange?: FieldHook[]
    afterRead?: FieldHook[]
    beforeChange?: FieldHook[]
    beforeValidate?: FieldHook[]
  }
  index?: boolean
  label?: Record<string, string> | false | string
  localized?: boolean
  name: string
  required?: boolean
  saveToJWT?: boolean | string
  unique?: boolean
  validate?: Validate
}
// ...
export type TextField = FieldBase & {
  admin?: Admin & {
    autoComplete?: string
    components?: {
      Error?: React.ComponentType<ErrorProps>
      Label?: React.ComponentType<LabelProps>
      afterInput?: React.ComponentType<any>[]
      beforeInput?: React.ComponentType<any>[]
    }
    placeholder?: Record<string, string> | string
    rtl?: boolean
  }
  maxLength?: number
  minLength?: number
  type: 'text'
} & (
    | {
        /** Makes this field an ordered array of strings instead of just a single strings. */
        hasMany: true
        /** Maximum number of strings in the strings array, if `hasMany` is set to true. */
        maxRows?: number
        /** Minimum number of strings in the strings array, if `hasMany` is set to true. */
        minRows?: number
      }
    | {
        /** Makes this field an ordered array of strings instead of just a single string. */
        hasMany?: false | undefined
        /** Maximum number of v in the strings array, if `hasMany` is set to true. */
        maxRows?: undefined
        /** Minimum number of strings in the strings array, if `hasMany` is set to true. */
        minRows?: undefined
      }
  )
// ...
export type Field =
  | ArrayField
  | BlockField
  | CheckboxField
  | CodeField
  | CollapsibleField
  | DateField
  | EmailField
  | GroupField
  | JSONField
  | NumberField
  | PointField
  | RadioField
  | RelationshipField
  | RichTextField
  | RowField
  | SelectField
  | TabsField
  | TextField
  | TextareaField
  | UIField
  | UploadField